Moved useResizeObserver hook to utils
